+++ code/programs/haskell/cowsay/test/Spec.hs
@@ -1,0 +1,6 @@
+import Test.Hspec
+
+import CowsaySpec (spec)
+
+main :: IO ()
+main = hspec spec